requirements.txt
coverage==5.5
deprecation==2.1.0
eyed3==0.9.7
filetype==1.2.0
packaging==23.0
python-dotenv==1.0.0
toml==0.10.2
